+++ vizro-core/examples/kpi/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Utils folder to contain helper functions and custom charts/components."""