+++ vizro-core/src/vizro/models/_components/ag_grid.py
@@ -105,8 +105,8 @@ class AgGrid(VizroBaseModel):
         )
 
         return dcc.Loading(
-            [
-                html.H3(self.title, className="table-title") if self.title else None,
+            children=[
+                html.H3(self.title) if self.title else None,
                 # The pagination setting (and potentially others) of the initially built AgGrid (in the build method
                 # here) must have the same setting as the object that is built by the on-page-load mechanism using
                 # with the user settings and rendered finally. Otherwise the grid is not rendered correctly.
@@ -117,4 +117,5 @@ class AgGrid(VizroBaseModel):
             ],
             color="grey",
             parent_className="loading-container",
+            overlay_style={"visibility": "visible", "opacity": 0.3},
         )