+++ vizro-core/src/vizro/models/_components/graph.py
@@ -136,6 +136,7 @@ class Graph(VizroBaseModel):
             ),
             color="grey",
             parent_className="loading-container",
+            overlay_style={"visibility": "visible", "opacity": 0.3},
         )
 
     @staticmethod