+++ vizro-core/tests/integration/test_examples.py
@@ -47,9 +47,10 @@ examples_path = Path(__file__).parents[2] / "examples"
     [
         (examples_path / "_dev", ""),
         (examples_path / "features", ""),
-        # (examples_path / "demo", ""),
-        # (examples_path / "_dev", "yaml_version"),
-        # (examples_path / "features", "yaml_version"),
+        (examples_path / "demo", ""),
+        (examples_path / "kpi", ""),
+        (examples_path / "_dev", "yaml_version"),
+        (examples_path / "features", "yaml_version"),
     ],
 )
 def test_dashboard(dash_duo, example_path, dashboard, version):