+++ vizro-core/tests/unit/vizro/conftest.py
@@ -43,7 +43,7 @@ def ag_grid_with_id(gapminder):
 
 @pytest.fixture
 def ag_grid_with_id_and_conf(gapminder):
-    return dash_ag_grid(id="underlying_ag_grid_id", data_frame=gapminder, dashGridOptions={"pagination": True})
+    return dash_ag_grid(id="underlying_ag_grid_id", data_frame=gapminder)
 
 
 @pytest.fixture